# backend/docs/swagger.yaml
basePath: /v1
definitions:
...
  analogdb.Color:
    properties:
      css:
        example: dimgray
        type: string
      hex:
        example: '#837d5c'
        type: string
      html:
        example: gray
        type: string
      percent:
        example: 0.55
        type: number
    type: object
...
  analogdb.Error:
    properties:
      code:
        example: not_found
        type: string
      message:
        example: not found
        type: string
    type: object
...
  analogdb.Image:
    properties:
      height:
        example: 7600
        type: integer
      resolution:
        example: high
        type: string
      url:
        example: https://cloudfront.net/1.jpeg
        type: string
      width:
        example: 4800
        type: integer
    type: object
  analogdb.Keyword:
    properties:
      weight:
        example: 0.43
        type: number
      word:
        example: baseball
        type: string
    type: object
...
  analogdb.Post:
    properties:
      aperture:
        example: f/2.0
        type: string
      author:
        example: thecameraman
        type: string
      camera_make:
        example: nikon
        type: string
      camera_model:
        example: fm2
        type: string
      caption:
        example: A woman on a beach at sunset
        type: string
      colors:
        items:
          $ref: '#/definitions/analogdb.Color'
        type: array
      created:
        description: time the post was added to analogdb
        example: "2025-07-11T12:00:00Z"
        type: string
      description:
        example: My favorite camera and film combo on 35mm at f/2.0
        type: string
      film_make:
        example: kodak
        type: string
      film_speed:
        example: 400
        type: integer
      film_type:
        example: color
        type: string
      focal_length:
        example: 35
        type: integer
      grayscale:
        example: false
        type: boolean
      id:
        example: 1
        type: integer
      images:
        items:
          $ref: '#/definitions/analogdb.Image'
        type: array
      keywords:
        items:
          $ref: '#/definitions/analogdb.Keyword'
        type: array
      nsfw:
        example: false
        type: boolean
      permalink:
        example: https://www.reddit.com/r/analog/comments/1/post
        type: string
      score:
        example: 1000
        type: integer
      sprocket:
        example: false
        type: boolean
      timestamp:
        example: 1752354541
        type: integer
      title:
        example: A day at the fields [Nikon FM2 | Portra 400]
        type: string
      updated:
        description: time the post was last changed in analogdb
        example: "2025-07-11T12:00:00Z"
        type: string
    type: object
...
  server.SearchMeta:
    properties:
      next_cursor:
        description: Opaque cursor for the next page, empty at the end
        example: eyJvIjo1MH0
        type: string
      next_page_url:
        example: /search?cursor=eyJvIjo1MH0&page_size=50&q=girl+at+sunset
        type: string
      page_size:
        example: 50
        type: integer
    type: object
  server.SearchResponse:
    properties:
      meta:
        $ref: '#/definitions/server.SearchMeta'
      posts:
        items:
          $ref: '#/definitions/analogdb.Post'
        type: array
      related_keywords:
        example:
        - beach
        - silhouette
        - dusk
        items:
          type: string
        type: array
    type: object
  server.SimilarPostsResponse:
    properties:
      posts:
        items:
          $ref: '#/definitions/analogdb.Post'
        type: array
    type: object
...
  server.encodePostsRequest:
    properties:
      batch_size:
        example: 20
        type: integer
      ids:
        items:
          type: integer
        type: array
    type: object
  server.encodePostsResponse:
    properties:
      failed_ids:
        items:
          type: integer
        type: array
      message:
        example: successfully encoded 5 posts
        type: string
    type: object
host: api.analogdb.com
info:
  contact: {}
  description: API for analogdb film database
  title: AnalogDB API
  version: "1.0"
paths:
...
  /encode:
    put:
      consumes:
      - application/json
      description: Encode posts to generate embeddings for similarity search. Supports
        both single post and batch encoding (requires authentication)
      parameters:
      - description: Post IDs and batch size for encoding
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.encodePostsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.encodePostsResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/analogdb.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/analogdb.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/analogdb.Error'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      security:
      - BasicAuth: []
      summary: Encode posts for similarity matching
      tags:
      - similarity
...
  /post/{id}/similar:
    get:
      consumes:
      - application/json
      description: Find posts similar to a given post using similarity matching
      parameters:
      - description: Post ID to find similar posts for
        in: path
        name: id
        required: true
        type: integer
      - default: 12
        description: Maximum number of similar posts to return
        in: query
        name: page_size
        type: integer
      - description: Include nsfw posts in query
        in: query
        name: nsfw
        type: boolean
      - description: Include b&w posts in query
        in: query
        name: grayscale
        type: boolean
      - description: Include sprocketshot posts in query
        in: query
        name: sprocket
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.SimilarPostsResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/analogdb.Error'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      summary: Find similar posts
      tags:
      - post
...
  /search:
    get:
      description: Hybrid search over image vectors, tags, captions and titles. Related
        keywords are returned on the first page only
      parameters:
      - description: Search text, 1 to 200 characters
        in: query
        name: q
        required: true
        type: string
      - default: 50
        description: Number of posts per page
        in: query
        name: page_size
        type: integer
      - description: Opaque cursor from next_cursor for the next page
        in: query
        name: cursor
        type: string
      - description: Filter by NSFW (true=only, false=exclude)
        in: query
        name: nsfw
        type: boolean
      - description: Filter by black and white (true=only, false=exclude)
        in: query
        name: grayscale
        type: boolean
      - description: Filter by sprocketshots (true=only, false=exclude)
        in: query
        name: sprocket
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.SearchResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      summary: Search posts by text
      tags:
      - search
  /search/image:
    post:
      consumes:
      - multipart/form-data
      description: Find posts that look like an uploaded JPEG, PNG or WebP image of
        at most 5 MB. Returns one page
      parameters:
      - description: Image to search with
        in: formData
        name: image
        required: true
        type: file
      - default: 50
        description: Number of posts to return
        in: query
        name: page_size
        type: integer
      - description: Filter by NSFW (true=only, false=exclude)
        in: query
        name: nsfw
        type: boolean
      - description: Filter by black and white (true=only, false=exclude)
        in: query
        name: grayscale
        type: boolean
      - description: Filter by sprocketshots (true=only, false=exclude)
        in: query
        name: sprocket
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.SearchResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/analogdb.Error'
        "413":
          description: Image too large
          schema:
            $ref: '#/definitions/analogdb.Error'
        "415":
          description: Unsupported image type
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
        "503":
          description: Too many image searches
          schema:
            $ref: '#/definitions/analogdb.Error'
      summary: Search posts by image
      tags:
      - search
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"