# backend/server/similarity.go
package server

import (
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
)

const (
	encodePath             = "/encode"
	defaultEncodeBatchSize = 20
	maxEncodeBatchSize     = 100
	encodeWriteTimeout     = 15 * time.Minute
)

func (s *Server) mountSimilarityHandlers(r chi.Router) {
	r.Route(encodePath, func(r chi.Router) {
		r.With(s.require(roleScraper)).Put("/", s.encodePosts)
	})
}

type encodePostsRequest struct {
	Ids       []int `json:"ids" examples:"1,2,3,4,5"`
	BatchSize int   `json:"batch_size" example:"20"`
}

type encodePostsResponse struct {
	Message   string `json:"message" example:"successfully encoded 5 posts"`
	FailedIDs []int  `json:"failed_ids"`
}

// @Summary Encode posts for similarity matching
// @Description Encode posts to generate embeddings for similarity search. Supports both single post and batch encoding (requires authentication)
// @Tags similarity
// @Accept json
// @Produce json
// @Param request body encodePostsRequest true "Post IDs and batch size for encoding"
// @Success 200 {object} encodePostsResponse
// @Failure 400 {object} analogdb.Error "Invalid request body"
// @Failure 401 {object} analogdb.Error "Unauthorized"
// @Failure 403 {object} analogdb.Error "Forbidden"
// @Failure 422 {object} analogdb.Error "Unprocessable entity"
// @Failure 500 {object} analogdb.Error "Internal server error"
// @Security BasicAuth
// @Router /encode [put]
func (s *Server) encodePosts(w http.ResponseWriter, r *http.Request) {
	if err := http.NewResponseController(w).SetWriteDeadline(time.Now().Add(encodeWriteTimeout)); err != nil {
		s.logger.WarnContext(r.Context(), "Fail extend write deadline for encode", "error", err)
	}

	var request encodePostsRequest
	if err := s.decodeBody(w, r, &request, "error parsing ids or batch_size from request body"); err != nil {
		s.writeError(w, r, err)
		return
	}

	if len(request.Ids) == 0 {
		s.writeError(w, r, badRequest("must provide ids to encode"))
		return
	}
	batchSize := clampLimit(request.BatchSize, defaultEncodeBatchSize, 1, maxEncodeBatchSize)

	var message string
	failedIDs := []int{}

	// encode single post
	if len(request.Ids) == 1 {
		err := s.SimilarityService.EncodePost(r.Context(), request.Ids[0])
		if err != nil {
			s.writeError(w, r, err)
			return
		}
		message = "successfully encoded post"

	} else {
		// encode batch of posts
		failed, err := s.SimilarityService.BatchEncodePosts(r.Context(), request.Ids, batchSize)
		if err != nil {
			s.writeError(w, r, err)
			return
		}
		failedIDs = append(failedIDs, failed...)
		message = fmt.Sprintf("successfully encoded %d posts", len(request.Ids)-len(failedIDs))
	}

	response := encodePostsResponse{
		Message:   message,
		FailedIDs: failedIDs,
	}
	if err := encodeResponse(w, r, http.StatusOK, response); err != nil {
		s.writeError(w, r, err)
	}
}